+++ Editor/Drawers/InputConditionDrawer.cs
@@ -1,5 +1,4 @@
 using ColliderEventSystem;
-using ColliderEventSystem.Editor;
 using UnityEditor;
 using UnityEngine;
 
@@ -43,8 +42,6 @@ namespace ColliderEventSystem.Editor.Drawers
             SerializedProperty triggerOnProp = so.FindProperty("triggerOn");
             Rect triggerOnRect = cursor.NextField(triggerOnProp);
             if (draw) EditorGUI.PropertyField(triggerOnRect, triggerOnProp, new GUIContent("Trigger On"));
-
-            ConditionHintDrawer.Layout(ref cursor, so, draw);
 
             return cursor.ConsumedHeight;
         }